### 2019-03-14 #1319 @arfon
@openjournals/joss-eics - the submitting author suggested @arokem as the handling editor.
