### 2019-03-14 #1319 @briatte
@KateHyoung Would be happy to review (and learn how reviewing for JOSS works in practice, by the same occasion)!
