### 2019-03-14 #1322 @whedon
[REVIEW]: UTDEventData: An R package to access political event data
**Submitting author:** @KateHyoung (<a href="http://orcid.org/0000-0002-0744-5884">Hyoungah Kim</a>)
**Repository:** <a href="https://github.com/KateHyoung/UTDEventData" target ="_blank">https://github.com/KateHyoung/UTDEventData</a>
**Version:** v1.0.0
