### 2019-03-16 #1322 @briatte
Quick note re: API registration, the UTD server does not check for email address validity and will return an error (500) if provided an invalid email address. That's not something that can be blamed on the package, of course, but the UTD server admins might benefit from improving their API registration form.
