### 2019-03-16 #1322 @briatte
The error message shown when the API key is wrong could be improved:

```
DataTables(api_key="foobar")
[1] "{\"STATUS\": \"ERROR\", \"DATA\":\"(<TYPE 'EXCEPTIONS.VALUEERROR'>, VALUEERROR('INVALID API KEY',), <TRACEBACK OBJECT AT 0X7F5ACD310488>)\"}"
```
