### 2019-03-16 #1322 @briatte
Typo in vignette:

> but `"cline_phenix"` will return __noting__.
