### 2019-03-16 #1322 @briatte
Comment # 5 [replicating vignette `UTDEventData.html`]

Thinking about the API key, it is customary to allow its storage as an environment variable (see [`?options`](https://stat.ethz.ch/R-manual/R-devel/library/base/html/options.html)). Is that possible with this package? The vignette does not mention it.
