### 2019-03-16 #1322 @briatte
Comment # 6 [replicating vignette `UTDEventData.html`]

The vignette is very helpful (and the authors do a great job at documenting possible errors/traps, e.g. Windows memory issues), yet I believe it would benefit from being broken down in more digestible chunks, such as:

- an 'intro' vignette explaining the gist of the package, how to cite it, and further useful links
- one or two 'example' vignettes (the first one could combine Examples 1 and 2)
- a 'designing complex queries' vignette covering things like regex-ed queries
