### 2019-03-16 #1322 @briatte
Comment # 7 [replicating vignette `UTDEventData.html`]

I'm done replicating the vignette.

Thinking about the package and data more globally, I think I'd appreciate a mention, somewhere in the docs, of how the data relate to (and whether it can be articulated with) similar event data and related event nomenclatures, e.g. those that Phil Schrodt has worked (or is working) on.

Apologies for not being more specific here, as I have limited experience with event data -- if this comment is too vague to be addressed, I'll inquire a bit and reformulate.
