### 2019-03-16 #1322 @briatte
Comment # 8 [reviewing `paper.bib`]

Re: this review point,

> __References:__ Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?

It seems to me that IEEE papers always come with a DOI. Those are not in the bibliography. Same goes for the Schrodt paper.

As for the Althaus _et al._ dataset (BibTeX reference `cline`), is that kind of data that could get distributed via e.g. Zenodo (and get a DOI from there)?
