### 2019-03-16 #1322 @briatte
Comment # 9 [checking my unchecked review points]

> __Functionality:__ Have the functional claims of the software been confirmed?

Yup. Checking off.

> __Performance:__ If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)

I do not think this really applies here. Performance in this package boils down (mostly) to API query speed. While replicating the vignette, I found some queries to be slow-ish, but given the size of the data returned by some examples (250,000+ obs.), I'm fine with it.

Checking off.

> __Automated tests:__ Are there automated tests or manual steps described so that the function of the software can be verified?

Also checking off this one, given that the package cannot really be tested against the API without providing an API key.

Some very basic unit tests could be imagined for small parts of the code: perhaps the authors will want to introduce a few control flow checks (e.g. making sure the argument to `andList` is a list).

> __Community guidelines:__ Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Also checking off this one, as I'm not sure the authors really expect community contributions.

That's pretty much it for me: over to @andrewheiss.

Last,
