### 2019-03-21 #1322 @briatte
Dear @KateHyoung (I hope it's fine to address submitting authors in JOSS review threads?)

I posted a bunch of numbered comments in the thread above. I hope that you will find them useful in improving your package, which I found helpful and carefully coded. Thanks for your work, looking forward to discussing things further.

All the best~
