### 2019-03-22 9f0eea2
updated doi in paper.bib
