+++ paper.bib
@@ -4,31 +4,32 @@
   booktitle={Intelligence and Security Informatics (ISI), 2016 IEEE Conference on},
   pages={139--144},
   year={2016},
-  organization={IEEE}
+  organization={IEEE},
+  url={https://doi.org/10.1109/isi.2016.7745457}
 }
 
 @inproceedings{brandt2018,
   title={Distributed Framework for Political Event Coding in Real-Time},
   author={Salam, Sayeed and Brandt, Patrick T. and Holmes, Jennifer and Khan, Latifur},
   booktitle={2nd European Conference on Electrical Engineering & Computer Science (EECS), Bern, Switzerland},
-  year={2010},
-  organization={IEEE}
+  year={2018},
+  organization={EECS}
 }
 
-@Manual{icews, 
- title={ICEWS Coded Event Data}, 
+@Manual{icews,
+ title={ICEWS Coded Event Data},
  author={Boschee, Elizabeth and Lautenschlager, Jennifer and O'Brien, Sean and Shellman, Steve and Starz, James and Ward, Michael},
  year={2015},
- note={Harvard Dataverse, V22}, 
+ note={Harvard Dataverse, V22},
  url={https://doi.org/10.7910/DVN/28075},
 }
 
-@Manual{cline, 
- title={Cline Center Historical Phoenix Event Data. v.1.0.0.}, 
+@Manual{cline,
+ title={Cline Center Historical Phoenix Event Data. v.1.0.0.},
  author={Althaus, Scott and Bajjalieh, Joseph and Carter, John F. and Peyton, Buddy and Shalmon, Dan  A.},
  year={2017},
  note={Distributed by Cline Center for Advanced Social Research.},
- url = {http://www.clinecenter.illinois.edu/data/event/}
+ url = {http://www.clinecenter.illinois.edu/data/event/phoenix/}
 }
 
 @inproceedings{vito2018,
@@ -37,7 +38,8 @@
   booktitle={2018 IEEE International Conference on Information Reuse and Integration (IRI)},
   pages={394--401},
   year={2018},
-  organization={IEEE}
+  organization={IEEE},
+  url={https://doi.org/10.1109/iri.2018.00065}
 }
 
 @article{schrodt2012,
@@ -48,7 +50,8 @@
   number={4},
   pages={546--569},
   year={2012},
-  publisher={Taylor \& Francis}
+  publisher={Taylor \& Francis},
+  url={https://doi.org/10.1080/03050629.2012.697430}
 }
 
 @inproceedings{halterman2017,
@@ -57,5 +60,6 @@
   booktitle={2017 IEEE International Conference on Big Data (Big Data)},
   pages={2879--2883},
   year={2017},
-  organization={IEEE}
+  organization={IEEE},
+  url={https://doi.org/10.1109/bigdata.2017.8258256}
 }
