### 2019-03-22 7f76c78
fixed the code of tableVar() to provide variable lists without whitespace
