+++ inst/doc/UTDEventData.Rmd
@@ -72,9 +72,9 @@ The codes/functions in this documentation have been tested in the following vers
 
 ## Searching functions
 
-The search functions help users to explore 1) the data tables in the UTD API server and 2) the attributes (variables) in a particular table. Moreover, Users can view the 100 sample data of each data table so as to check its structure, particular values, and patterns of attributes before downloading them on your local machines.
+The search functions help users to explore 1) the data tables in the UTD API server and 2) the attributes (variables) in a particular table. Moreover, users can view the 100 sample data of each data table so as to check its structure, particular values, and patterns of attributes before downloading them on your local machines.
 
-`DataTables()` returns all data table names in the UTD server. 
+`DataTables()` returns a list of all data tables in the UTD server. 
 
 ```{r,eval = FALSE}
 # returning all data table the server contains with entering an API key
@@ -87,7 +87,7 @@ DataTables(k)
 "'PHOENIX_RT', 'CLINE_PHOENIX_SWB', 'CLINE_PHOENIX_FBIS', 'CLINE_PHOENIX_NYT', 'ICEWS'"
 ```
 
-`tableVar()` returns the list of variables (attributes) in a particular data table a user specified in the function. For example, the variables in the Phoenix real-time (RT) data table can be obtained as shown in the following example. 
+`tableVar()` returns a list of variables (attributes) in a particular data table a user specified in the function. For example, the variables in the Phoenix real-time (RT) data table can be obtained as follows. 
 
 ```{r, eval = FALSE}
 tableVar(api_key="...", table = "phoenix_rt")
@@ -101,9 +101,9 @@ tableVar(k, "Icews")
 tableVar(k, "Cline_Phoenix_swb")
 ```
 
-This function is __not__ case sensitive, so a user may type either lower case or upper case of the data table name. However, the full name as returned by `Table()` should be entered in the function when applying the table name in other functions. For instance, `"cline_Phoenix_swb"` and `"CLINE_PHOENIX_SWB"` will return the same variables, but `"cline_phenix"` will return noting. 
+This function is __not__ case sensitive, so users may type either lower case or upper case of a data table name. However, a text of the full name as returned by `Table()` should be entered in the function when applying the table name in other functions. For example, Inputs of `"cline_Phoenix_swb"` and `"CLINE_PHOENIX_SWB"` will return the same variables, but one of `"cline_phenix"` will return a message that ask users to check a table name again. 
 
-This function has another feature that a user can look up a particular variable in a data set. For example, one who may wonder the ICEWS data have a variable named "target" can type the string of the certain variable as follows;
+This function has another feature that a user can look up a particular variable users may want. For example, one who may wonder the ICEWS data have a variable named "target", can type the string of 'target' as follows;
 
 ```{r, eval = FALSE }
 tableVar(api_key = "...", table = "icews", lword = "target")
@@ -115,7 +115,7 @@ tableVar(k, table="icews", lword="target")
 ```
 
 
-`previewData()` shows the 100 observations of a data table as an example. Occasionally, users need to know a pattern of variables or a data structure to analyze the data in R formats. This function can be useful when building a query block with a regular expression (`returnRegExp()`) function. 
+`previewData()` shows 100 observations of a data table as an example. Occasionally, users need to know a pattern of variables or a data structure to analyze the data in R dataframes. This function can be useful when users want to see some pattenrs of an interesting variable to build a query block with a regular expression (`returnRegExp()`) function. 
 
 ```{r, eval = FALSE }
 dataSample <- previewData(api_key = " ", table_name = "PHOENIX_RT")
@@ -126,17 +126,17 @@ View(dataSample)
 
 ## Subsetting function I
 
[…]
