### 2019-03-22 62544dc
cleaning vignettes
