### 2019-03-24 c2c45b9
fixed an error of returnRegExp()
