+++ vignettes/UTDEventData.R
@@ -1,0 +1,277 @@
+## ---- eval = FALSE-------------------------------------------------------
+#  # install the package without the vignette
+#  devtools::install_github("KateHyoung/UTDEventData")
+#  
+#  # install the package with the vignette
+#  devtools::install_github("KateHyoung/UTDEventData", build_vignettes=TRUE)
+
+## ----eval = FALSE--------------------------------------------------------
+#  # returning all data table the server contains with entering an API key
+#  DataTables(api_key = " ")
+#  "'PHOENIX_RT', 'CLINE_PHOENIX_SWB', 'CLINE_PHOENIX_FBIS', 'CLINE_PHOENIX_NYT', 'ICEWS'"
+#  
+#  # suggesting a way to avoid repetitive typing an API key into functions
+#  k <-"...api key...."
+#  DataTables(k)
+#  "'PHOENIX_RT', 'CLINE_PHOENIX_SWB', 'CLINE_PHOENIX_FBIS', 'CLINE_PHOENIX_NYT', 'ICEWS'"
+
+## ---- eval = FALSE-------------------------------------------------------
+#  tableVar(api_key="...", table = "phoenix_rt")
+#  
+#  # an easy way of applying a stored API text to avoid the repeation of API key typing
+#  k <-"...api key...."
+#  tableVar(k, "Phoenix_rt")
+#  
+#  tableVar(k, "Icews")
+#  
+#  tableVar(k, "Cline_Phoenix_swb")
+
+## ---- eval = FALSE-------------------------------------------------------
+#  tableVar(api_key = "...", table = "icews", lword = "target")
+#  
+#  # when a user wants to know the attribute that labeled as 'target' in ICEWS
+#  k <- "..api key..."
+#  tableVar(k, table="icews", lword="target")
+#  '" Target Name"    " Target Sectors", ....'
+
+## ---- eval = FALSE-------------------------------------------------------
+#  dataSample <- previewData(api_key = " ", table_name = "PHOENIX_RT")
+#  View(dataSample)
+
+## ---- eval = FALSE-------------------------------------------------------
+#  ## several examples of different data tables with citation texts
+#  k <-'api key...'
+#  subset1 <- pullData(api_key = k, table_name = "phoenix_rt", country = list('canada','China'), start = '20171101',  end = '20171102', T)
+#  
+#  # to store the data only
+#  data <- subset1$data
+#  
+#  # to print the citation texts
+#  subset1$citation
+#  
+#  # to avoid the citation texts
+#  data <- pullData(k, "phoenix_rt", list('canada','China'), '20171101', '20171102',
+#           citation = FALSE)
+
+## ---- eval=FALSE---------------------------------------------------------
+#  # creating an object
+#  obj <- Table$new()
+#  
+#  # setting an object of an API key
+#  obj$setAPIKey("....")
+#  obj$DataTables()  # returns the available data tables in the UTD server
+#  obj$tableVar("cline_Phoenix_NYT")
+#  
+#  # when a user wants to subset real-time data ('phoenix_rt) from 20171101 to 20171102 on MEX(Mexico)
+#  obj$pullData("Phoenix_rt", list("MEX"),start="20171101", end="20171102")
+
+## ---- eval=FALSE---------------------------------------------------------
+#  # to estimate the data size of the entire Cline_Phoenix_NYT data
+#  getQuerySize(api_key = , table_name ='Cline_Phoenix_NYT', query = 'entire')
+#  
+#  # to download the data
[…]
