### 2019-03-26 a469ff9
Update SearchFunctions.R
Creating an error message when an invalid API key is entered
