### 2019-03-27 a64c036
Update DOIs for paper.bib
