+++ paper.bib
@@ -5,7 +5,8 @@
   pages={139--144},
   year={2016},
   organization={IEEE},
-  url={https://doi.org/10.1109/isi.2016.7745457}
+  url={https://doi.org/10.1109/isi.2016.7745457},
+  doi = {10.1109/isi.2016.7745457}
 }
 
 @inproceedings{brandt2018,
@@ -22,6 +23,7 @@
  year={2015},
  note={Harvard Dataverse, V22},
  url={https://doi.org/10.7910/DVN/28075},
+ doi = {10.7910/DVN/28075}
 }
 
 @Manual{cline,
@@ -39,7 +41,8 @@
   pages={394--401},
   year={2018},
   organization={IEEE},
-  url={https://doi.org/10.1109/iri.2018.00065}
+  url={https://doi.org/10.1109/iri.2018.00065},
+  doi={10.1109/iri.2018.00065}
 }
 
 @article{schrodt2012,
@@ -51,7 +54,8 @@
   pages={546--569},
   year={2012},
   publisher={Taylor \& Francis},
-  url={https://doi.org/10.1080/03050629.2012.697430}
+  url={https://doi.org/10.1080/03050629.2012.697430},
+  doi={10.1080/03050629.2012.697430}
 }
 
 @inproceedings{halterman2017,
@@ -61,5 +65,6 @@
   pages={2879--2883},
   year={2017},
   organization={IEEE},
-  url={https://doi.org/10.1109/bigdata.2017.8258256}
+  url={https://doi.org/10.1109/bigdata.2017.8258256},
+  doi={10.1109/bigdata.2017.8258256}
 }
