### 2019-03-27 #1322 @KateHyoung
Dear @briatte and other reviewers

I have made some changes corresponding to the comments from @briatte.
1.	Response for the comment #2
I have fixed the code of DataTables(). With an invalid API key, the function returns a message.
> DataTables("fooboo")
[1] "This API key is invalid. Please check that you have entered your key correctly."

Ref. https://github.com/KateHyoung/UTDEventData/commit/a469ff91d028897a2e7dd6804e434fb64ab90cc5

2.	Response for the comment #3
I have fixed the function of tableVar() that now prints a variable list without a space. Along with the change of this function, returnRexExp() was also fixed to provide appropriate query elements for API syntax.
https://github.com/KateHyoung/UTDEventData/commit/7f76c78f9c618a225acc5e8dbc12117732dd5a1d#diff-6134badaa82d09a6bcf0dffd6b2f97b4
> tableVar(k, "Phoenix_rt")
 [1] "code"            "src_actor"       "month"           "tgt_actor"

3.	Response for the comment #4
I am thinking which way is better for users. If I make it to accept country objects directly, users should provide it as c(“ctr”, “time”) in order to generate a proper query syntax. I feel that listing query elements with list() is no big difference with accepting the query element with c().  But if I have a better idea for it. I will update it.
4.	Response for the comment #8
I have updated the DOIs of the reference in a paper.
Ref.
https://github.com/KateHyoung/UTDEventData/commit/a64c03691cc8dd50e5b36cd420e85c161cba4328#diff-68dc399b92868c40cec2beeb0a59b48e
https://github.com/KateHyoung/UTDEventData/commit/9f0eea2b4fd6f6a6717670349e010fadfa1274d3#diff-68dc399b92868c40cec2beeb0a59b48e

I am working on editing the vignette according to the comments on it. Once I have done for this work I will post it soon.
Thank you again,
Kate
