+++ vignettes/UTDEventData.Rmd
@@ -53,7 +53,7 @@ The UTD Event Data server has the five different event data tables.
 |Cline Phoenix NYT  | 1945 - 2005       | [Cline Center](http://www.clinecenter.illinois.edu/data/event/phoenix/)|
 |Cline Phoenix FBIS | 1945 - 2005       | [Cline Center](http://www.clinecenter.illinois.edu/data/event/phoenix/)|
 |Cline Phoenix SWB  | 1979 - 2015       | [Cline Center](http://www.clinecenter.illinois.edu/data/event/phoenix/)|
-|TERRIER            | 1979 - 2016        | [TERRIERDATA.org](http://terrierdata.org)|
+|TERRIER            | 1979 - 2016       | [TERRIERDATA.org](http://terrierdata.org)|
 
 
 You can find the information of each data table in the specified websites such as the contents of data, their entities, and attributes. 
