### 2019-03-27 87f4281
updated orList() and andList() functions
