+++ vignettes/UTDEventData.Rmd
@@ -28,8 +28,10 @@ vignette: >
 
 # Introduction
 
-This R package allows users to extract a data set from the API Event Data server at the University of Texas at Dallas (UTD). The project of the UTDEventData R package is in progress and has been updated day by day. Your comments, feedback, and suggestions are welcome so that more user friendly methods are prepared in the library.
-If you have questions in using the package, please contact Kate Kim (<hyoungah.kim@utdallas.edu>) at UTD.  
+The R library, `UTDEventData`, provides the direct access to the political event database at the University of Texas at Dallas (UTD). Using this library, researchers can query the event data stored at the UTD's API server and load them into R for subsequent analysis. Several functions are prepared in the library to extract and to analize the data according to users' preference. This library also provides the citation functions ont only for data tables but also the library itself. 
+
+The project of the UTDEventData R package is in progress and has been updated day by day. Your comments, feedback, and suggestions are welcome so that more user friendly functions are prepared in the library.
+If you have questions in using the package, please contact Kate Kim (<hyoungah.kim@utdallas.edu>).  
 
 This package requires you to have an API key to access to the UTD data server. Please find the following link and fill out the form to obtain an API key: <http://eventdata.utdallas.edu/signup>. 
 
@@ -42,8 +44,18 @@ devtools::install_github("KateHyoung/UTDEventData")
 devtools::install_github("KateHyoung/UTDEventData", build_vignettes=TRUE)
 ```
 
+To cite the package, an excution function `citation()` can be used, and users can find the citatation texts when using the data obtaining functions, `pullData()` and `sendQuery()` with a defualt option of `citation = TRUE`. Moreover, the function of `citeData()` in the package will print data ciatation texts in a R console. 
+```{r, eval = FALSE}
+# to cite the package 
+citation("UTDEventData")
 
-The UTD Event Data server has the five different event data tables. 
+# to cite a data table
+citeData(table_name = "Data table name")
+```
+
+
+The UTD API server contains the six different data tables; Real-time Phoenix, ICEWS, three other Cline's Phoenix data sets, and TERRIER. With the several searching and extracting methods in this library, political and social scientists can explore the contents of each data table and can obtain the historic and real-time data for their research.  We expect that this R library provides the better evivornment in accessing to the event data for R users, so the number of studies on the large scale event data will increase in this drawing of the big data age.
+
 
     Table 1:  Data Table Information       
 |Data Table         | Timeline          |  Further Information|
@@ -158,10 +170,27 @@ data <- pullData(k, "phoenix_rt", list('canada','China'), '20171101', '20171102'
 As illustrated, data are able to be retrieved by the country names and time ranges a user specified. A country in a data set means the country where a particular event happened. This country might not be the same with the source or target countries of an event. A researcher should pay attention to this attribute when conducting research. This feature of each variable can be scrutinized by exploring the variables of a sample data downloaded by `tableVar()` or `previewData()`. 
 
 
-### Reference class for appying an API key to functions 
+### Handing an API key
 
[…]
