### 2019-04-05 e41ee52
updated package docs
