+++ inst/doc/UTDEventData.R
@@ -5,6 +5,13 @@
 #  # install the package with the vignette
 #  devtools::install_github("KateHyoung/UTDEventData", build_vignettes=TRUE)
 
+## ---- eval = FALSE-------------------------------------------------------
+#  # to cite the package
+#  citation("UTDEventData")
+#  
+#  # to cite a data table
+#  citeData(table_name = "Data table name")
+
 ## ----eval = FALSE--------------------------------------------------------
 #  # returning all data table the server contains with entering an API key
 #  DataTables(api_key = " ")
@@ -54,6 +61,14 @@
 #           citation = FALSE)
 
 ## ---- eval=FALSE---------------------------------------------------------
+#  k <- '...your API key....'
+#  DataTables(k)
+
+## ---- eval=FALSE---------------------------------------------------------
+#  Sys.setenv(APIKEY = "...your API key...")
+#  DataTables(Sys.getenv("APIKEY"))
+
+## ---- eval=FALSE---------------------------------------------------------
 #  # creating an object
 #  obj <- Table$new()
 #  
@@ -75,7 +90,7 @@
 
 ## ---- eval = FALSE-------------------------------------------------------
 #  # basic usage
-#  sendQuery(api_key='', tabl_name ='', query = list(), citation = TRUE)
+#  sendQuery(api_key='', tabl_name ='', query = c(), citation = TRUE)
 #  
 #  # to store the ICEWS subset in the vector of myData without the citation
 #  # query_block is a list of the quries built by the query block functions illustrated in subchapters
@@ -112,50 +127,10 @@
 
 ## ---- eval = FALSE-------------------------------------------------------
 #  # combine stored query blocks such as 'time' or 'locQuery' created before
-#  and_query <- andList(query_prep = list(locQuery, time))
+#  and_query <- andList(query_prep = c(locQuery, time))
 #  
 #  # subset with two or more stored query blocks such as 'locQuery' or 'dyad'
-#  or_query <- orList(query_prep = list(locQuery, dyad))
-
-## ---- eval=FALSE---------------------------------------------------------
-#  
-#  # examples of subsetting functions II
-#  
-#  # creating query blocks
-#  # a country constrain of 'CHN' and 'USA'
-#  k <- 'api_key'
-#  ctr <- returnCountries("phoenix_rt", list("CHN", "USA"))
-#  
-#  # A query of time between 2017-11-1 and 2017-11-5
-#  time <- returnTimes("phoenix_rt", "20171101","20171105")
-#  
-#  # a boolean logic, or, with the two query blocks
-#  or_query <- orList(list(ctr, time))
-#  # request a data set to the API server with the package citation
-#  d1 <- sendQuery(k,"phoenix_rt",or_query, TRUE)
-#  
-#  # to view the subset
-#  head(d1$data, 10)
-#  View(d1)
-#  
-#  # a boolean logic, and, with the two query blocks
-#  and_query <- andList(list(ctr, time))
-#  d2 <- sendQuery(k, "phoenix_rt", and_query, TRUE)
-#  
-#  # to view the subset
-#  head(d2$data, 10)
-#  View(d2)
-#  
-#  # when a user wants to extract all event in US and China with the events for which the source was a government actor from the Phoenix real-time data
-#  rgex <- returnRegExp(k, "phoenix_rt","GOV", "src_agent")
[…]
