### 2019-04-07 #1322 @alexhanna
Thanks so much for the comprehensive comments, @briatte, and your responses @KateHyoung.

@andrewheiss have you had a chance to look through the submission?
