### 2019-04-08 6e2cd05
Fix some minor typos
