+++ vignettes/UTDEventData.Rmd
@@ -68,7 +68,7 @@ citeData(table_name = "Data table name")
 ```
 
 
-## Computer Enviroments This Package Has Tested
+## Computer Environments This Package Has Tested
 The codes/functions in this documentation have been tested in the following versions of:
 
   - Microsoft Open R 3.4.0 
@@ -125,7 +125,7 @@ tableVar(k, table="icews", lword="target")
 ```
 
 
-`previewData()` shows 100 observations of a data table as an example. Occasionally, users need to know a pattern of variables or a data structure to analyze the data in R dataframes. This function can be useful when users want to see some pattenrs of an interesting variable to build a query block with a regular expression (`returnRegExp()`) function. 
+`previewData()` shows 100 observations of a data table as an example. Occasionally, users need to know a pattern of variables or a data structure to analyze the data in R dataframes. This function can be useful when users want to see some patterns of an interesting variable to build a query block with a regular expression (`returnRegExp()`) function. 
 
 ```{r, eval = FALSE }
 dataSample <- previewData(api_key = " ", table_name = "PHOENIX_RT")
@@ -144,7 +144,7 @@ An input format of dates is "YYYYMMDD", and a time range should be within a time
 
 The country indicates the geolocation of events such as the places in which a particular event happened in a news article. The country inputs can be either full names or the ISO-Alpha3 code. In this inputs, please use a consistent format when you type countries names. We recommend the ISO-3 code format in order to reduce systemic errors in the function.
 
-The function returns citation texts \& BibTeX for publication at the end of data retrieval. To avoid printing them in your R-console, please turn off the `citation` option by logiscally choosing "FALSE" in the function. 
+The function returns citation texts \& BibTeX for publication at the end of data retrieval. To avoid printing them in your R-console, please turn off the `citation` option by logically choosing "FALSE" in the function. 
 
 The following code is an example usage of `pullData()`. 
 
@@ -170,7 +170,7 @@ As illustrated, data are able to be retrieved by the country names and time rang
 
 ### Handing an API key
 
-Throwing an API key into functions evvery time could be annoying tasks for users. To avoid troublesome work, we suggest some convenient approaches. 
+Throwing an API key into functions every time could be annoying tasks for users. To avoid troublesome work, we suggest some convenient approaches. 
 
 #### Store an API key in a R memory
 This approach is to simply store an API key in a global environment of R and call it when you want to use it. Although it is simple, the API key should be re-entered when a R is re-run. 
@@ -180,7 +180,7 @@ DataTables(k)
 ```
 
 #### Store an API key as an environment variable 
-This option is to declare user's API key as an environemtn variable and use it by `Sys.getenv()`. This option helps users to avoid a hassle to input the key each time. 
+This option is to declare user's API key as an environment variable and use it by `Sys.getenv()`. This option helps users to avoid a hassle to input the key each time. 
 ```{r, eval=FALSE}
 Sys.setenv(APIKEY = "...your API key...")
 DataTables(Sys.getenv("APIKEY"))
@@ -188,7 +188,7 @@ DataTables(Sys.getenv("APIKEY"))
 
 
 #### Using a reference class
-Theird option is using a reference class, named `Table()` , prepared in this library. One downsiade of this is that the function works only with `DataTables()`, `tableVar()`, and `pullData()`. 
+Third option is using a reference class, named `Table()` , prepared in this library. One downside of this is that the function works only with `DataTables()`, `tableVar()`, and `pullData()`. 
 ```{r, eval=FALSE}
 # creating an object
 obj <- Table$new()
@@ -203,9 +203,9 @@ obj$pullData("Phoenix_rt", list("MEX"),start="20171101", end="20171102")
 ```
 
 ### Extracting the entire data for a particular data set
[…]
