### 2019-04-08 KateHyoung/UTDEventData#10 @andrewheiss
The Table class isn't working
Following the vignette, creating an object from the `Table` class doesn't work:

```r
# creating an object
obj <- Table$new()
# Error: object 'Table' not found
```
