### 2019-04-08 #1322 @andrewheiss
**UTDEventData** is a fantastic new package that provides easy access to [UT Dallas's event data server](http://eventdata.utdallas.edu/). This is an important package, given the complexities involved in pulling event data from half a dozen different sources, and I've already started to use it in my own research. The package is easy to install and use, and I only came across a few issues, noted below. I've opened issues for a few of these at the package repository.

## Documentation

The README is fairly sparse and doesn't include any details on how to use the package (beyond providing a list of possible functions). The vignette currently contains complete examples of how to use the package, but reading those examples outside of R (i.e. through GitHub) is trickier since GitHub doesn't display HTML.

The README should include instructions about how to set up and use the API key, either as a local variable (e.g. `k <- "blah"`) or as a system variable (e.g. adding `Sys.setenv(APIKEY = "blah")` to `.Rprofile`).

I have not really ever come across instructions to build vignettes when building packages with **devtools**, even when a package has vignettes ([see **gganimate**, for example](https://github.com/thomasp85/gganimate)), and most packages I've seen on GitHub don't include two different options for package installation (except when one version is on CRAN and one is on GitHub). Additionally, nothing in the vignette is actually run when the package is built, since `eval=FALSE` is enabled on all the vignette chunks. This makes sense to prevent overloading the event data server, and it makes it so users don't need an API key to install the package, but if nothing in the vignette actually runs, it might be easier to include lots of the documentation and examples in the README so users (1) don't have to explicitly build vignettes to see how the package works, and (2) users can follow rendered HTML online through GitHub.

But there's also utility in actually using vignettes so that users can run `vignette("UTDEventData")` within R to see documentation.

There's no right answer here. I don't have any official suggestion for README vs. vignettes, other than the fact that the README should at least include some quick "getting started right away" examples.

In a similar vein, the practical examples at the end of the vignette are extremely helpful, but they're buried at the bottom of the long vignette. Perhaps consider splitting the vignette into two parts—one highlighting the different ways of using the different functions, and one with the examples so that users can jump to actual usage faster.

## API access

Allowing users to store the API key in an environment variable is convenient. It might be worth changing the name of the system variable. Right now it is `APIKEY`, but it's hard to distinguish that from other system-level variables users might be using. Adding a prefix like `UTD_APIKEY` or `UTD_API` would help with recognizability.

Additionally, it might be useful (someday later, perhaps—not necessarily for this release) to have the system-level API key a default option in functions like `DataTables()` or `tableVar()` so that users can run `tableVar("Phoenix_rt")` without needing to pass the `api_key` argument every time. Other functions like `devtools::install_github()` do something similar—if there's a `GITHUB_PAT` system variable, it uses it, otherwise it doesn't.

## Functionality issues

The biggest issue is that the S4 class doesn't seem to be working correctly right now. When I run `obj <- Table$new()`, I get the following error:

```r
# > obj <- Table$new()
# Error: object 'Table' not found
```

As such, I'm unable to test the rest of the `obj$*` commands. (see https://github.com/KateHyoung/UTDEventData/issues/10)

I see the utility of building the API key into the `Table` objects, but the documentation doesn't explain why this method is preferable to the other more documented approaches to extracting data. The vignette mentions that using `Table` objects simplifies API calls, but that it is limited to just `DataTables()`, `tableVar()`, and `pullData()`. What should users do if they want to use other functions like `sendQuery()`? Is there value in maintaining two different ways of accessing data (especially when one method is incomplete)? (I'm assuming there is value—it might be helpful to explain the difference in approaches).

The "Data request function" section doesn't reproduce and there's minimal explanation of what `query_block` should look like. It says to pass "a list of queries created by `andList()`, `orList()`, or a single query block in the function as shown in the following example code," but the documentation hasn't covered `andList()` and `orList()` yet, and there's no working single query block in the example code. For now, running the example `sendQuery()` as it stands in the vignette produces an error:

```r
myData <- sendQuery(k,"icews", query_block, citation = TRUE)

# Error in sendQuery(k, "icews", query_block, citation = TRUE) :
#   object 'query_block' not found
```

It might be helpful to include code for creating an example `query_block` object that users can use before delving into all the query examples that follow.

There's a comment in the first example that "`# the data set d1 and d2 are identical`", but there is no `d1` or `d2`. I'm guessing that these are now `dt1` and `dt`, but maybe they should be renamed?

The call to `xtable()` in example 2 assumes that the user has loaded the xtable library already, but that never happens explicitly in the vignette. It might be helpful to prefix the function with `xtable::xtable()`. Alternatively, since you're using rmarkdown to generate the vignette, consider using `pander::pandoc.table()` or `knitr::kable()` to generate a Markdown table instead of raw TeX.

## Clarity and quality of writing

The vignette is generally easy to follow and is a helpful guide to using the package.

There are several typos and grammatical errors in the vignette. I've corrected some of the typos in https://github.com/KateHyoung/UTDEventData/pull/9 - a quick round of editing should clean up the rest.

## Testing

There are no formal tests, likely because the package is heavily reliant on remote access to an API and it would be computationally costly to hit the server for each test. Creating a cache of data locally might be helpful for testing, or perhaps adding tests for non-server-based functions like `returnLatLon()` or `returnDyad()` could be helpful (this isn't necessary, though)

## Contributions

It might be helpful to have some community guidelines in the package too, such as a CONTRIBUTING.md file (perhaps modeled after something like [this](https://github.com/ropensci/nomisr/blob/master/CONTRIBUTING.md) or [this](https://github.com/tidyverse/ggplot2/blob/master/CONTRIBUTING.md)) and a CONDUCT.md file (like [this](https://github.com/ropensci/qualtRics/blob/master/CONDUCT.md) or [this](https://github.com/tidyverse/readr/blob/master/CONDUCT.md))

----

Phew. That's all I've got. I'm excited for this package!
