+++ NAMESPACE
@@ -1,6 +1,7 @@
 # Generated by roxygen2: do not edit by hand
 
 export(DataTables)
+export(Table)
 export(andList)
 export(citeData)
 export(entireData)
