### 2019-04-09 KateHyoung/UTDEventData#10 @KateHyoung
Thank you, @andrewheiss, for finding the error of Table.
I have fixed it, and it is working now as no error returns.
``` r
> obj<-Table$new()
```
Nevertheless, our team is thinking to drop out this reference class from the library because we think that storing an API key as an environment variable is more convenient.

Please let us know if you have a better idea on this issue.
