### 2019-04-09 KateHyoung/UTDEventData#10 @andrewheiss
Hi! I think the best solution is really to use environment variables, as you've started doing, but wire them into the functions more directly instead of requiring `api_key` in every function. Packages like [**analogsea**](https://github.com/sckott/analogsea) and [**qualtRics**](https://github.com/ropensci/qualtRics) do this really well, storing API credentials as environment variable in `~/.Rprofile` and not requiring them to be passed through functions, while also avoiding complex S4-style classes
