### 2019-04-13 432c713
adding a prefix of API KEY, editing a vignette, updating README.md, and new CONDUCT.md
