### 2019-04-13 4327c12
minor correction in README.md
