### 2019-04-17 84d04de
README.md wording / API key default
If utd_api_key is not passed, attempt to load it from the environment variable UTDAPIKEY. Please consider moving the utd_api_key to the last argument, so that unnamed arguments may be passed, without passing NA first.
