+++ README.md
@@ -1,11 +1,12 @@
+
 # UTDEventData ver. 1.0.0
 
-This R package helps users to extract data from the UTD Event Data server. The project of the UTDEventData R package has not fully completed and is daily updated. Your comments, feedback, and suggestions are welcome to enhence the package.   
-If you have any question regarding the package, please contact Kate Kim (<hyoungah.kim@utdallas.edu>).
+The UTDEventData R package provides an interface to extract data from the UTD Event Data server. This package is stable and actively maintained/updated. Your comments, feedback and suggestions are welcome.   
+If you have any question regarding the package, please contact Kate Kim (<hyoungah.kim@utdallas.edu>), or open an issue (<https://github.com/KateHyoung/UTDEventData/issues>).
 
-This R package is part of a project, titled "Modernizing Political Event Data for Big Data Social Science Research". More information can be found on [the project webpage](http://eventdata.utdallas.edu/data.html).
+This package is part of the "Modernizing Political Event Data for Big Data Social Science Research" project. More information can be found on [the project webpage](http://eventdata.utdallas.edu/data.html).
 
-Several functions to explore and extract data in this package are listed below. More details of these methods are illustrated in the vignette of this package. 
+Several functions to preview and download data are listed below. More details of these methods are illustrated in the vignette. 
 
 - Table: a reference class 
 - citeData( ): for citing the package and data tables in the UTD server for publications
@@ -18,73 +19,76 @@ Several functions to explore and extract data in this package are listed below. 
 - sendQuery( ): for requesting built queries from the API server to download data
 
 ****
-The Query Block functions:
+Leaf Query Block functions:
+- returnTimes( ): create a query block by time periods
+- returnCountries( ): create a query block by countries
+- returnLatLon( ): create a query block by latitude and longitude
+- returnDyad( ): create a query block of a dyad for both source and target actors
+- returnRegExp( ): create a query block by pattern of attributes in a data table
 
-- retrunTimes( ): creating a query block by time periods
-- returnCountries( ): creating a query block by countries
-- returnLatLon( ): creating a query block of region(s) with latitude and longitude
-- returnDyad( ): creating a dyad query of both source and target actors
-- returnRegExp( ): creating a query block by pattern of attributes in a data table
-- orList( ): creating a union of query blocks 
-- andList( ): creating an intersection of query blocks
-
-This package requires you to have an API key to access to the UTD data server. Please find the following link and fill out the form to obtain an API key. 
-<http://eventdata.utdallas.edu/signup>
+Branch Query Block functions:
+- orList( ): match records that satisfy any of the child query blocks 
+- andList( ): match records that satisfy all of the child query blocks 
 
 ## Installation
 
+Without the vignette:
 ```
-# without the vignette
 devtools::install_github("KateHyoung/UTDEventData") 
+```
 
-# with the vignette
+With the vignette:
+```
 devtools::install_github("KateHyoung/UTDEventData", build_vignettes=TRUE)
 ```
-## UTD API key
-This package requires users to have a UTD API key to access to the UTD data server. Please find the following link and fill out the form to obtain an API key: <http://eventdata.utdallas.edu/signup>.
 
-### Handling an API key
-#### Store an API key in a R memory
[…]
