### 2019-04-17 3225b7e
Updated API key options in functions
API key storing as envs. embeding in some functions
clearify README and vignette files
