### 2019-04-17 ddde730
updated package docs corresponding to code updated
