### 2019-04-17 #1322 @KateHyoung
Dear reviewers,

Thank you for your helpful comments for our package. I thoughtfully went over every comment and tried to reflect all of them because those make sense and enhance the package’s operations.

For the @briatte ‘s comment #4, we have changed the function code as you suggested. Users now can throw objects directly rather than list up with `list()`.
Ref. https://github.com/KateHyoung/UTDEventData/commit/87f42817ba1e17734cffea4559130c734738ec5f#diff-079497ecdda4a05bbcdba43d0d5cacac

For the @briatte ‘s comment #7, I have added paragraphs to describe how this package is linked to the other previous political event studies in the introduction part of the vignette.
https://github.com/KateHyoung/UTDEventData/commit/e41ee52f4d24d79435d57091ea64d34495135a17#diff-a28523db019dca3cb3eaab6cda57fe4d

@andrewheiss, thank you for the idea of handling API access in the package. I have learned lots of things from the information you provided me. I have incorporated the way to set an API key in some functions as an environment variable. See the updates in `DataTables()`, `tableVar()`, and `prevewData()` (Please see README and vignette files). I have tried to apply the method to apply to other functions that require an API key and noticed that some functions created minor issues because of the argument setting as breaking a function. I need more time to make them stable working without errors, so I leave it for later works and will complete it during a summer break. I hope this work is good enough for this version.
Also, I have added a prefix to `api_key` variable as to be `utd_api_key`.
Ref. https://github.com/KateHyoung/UTDEventData/commit/4db04487d88a66e43fabe2ec67c478413bc5f04f

For the comments in Functionality issues, I have fixed the code so that `Table` is working well. Ref. https://github.com/KateHyoung/UTDEventData/commit/159b807815bc36bf34a74baeeef4b39e189c6c5a#diff-ada2ba640dbb79b8d3a115369830f718
```
> obj<-Table$new()
> obj$setAPIKey("utd_api_key")
> obj$DataTables()
[1] "'CLINE_PHOENIX_SWB', 'CLINE_PHOENIX_NYT', 'ICEWS', 'PHOENIX_RT', 'CLINE_PHOENIX_FBIS', 'TERRIER'"
```
I have changed/deleted some codes not completely works in R code chunks in the vignette.  And the `xtable()` code is also changed to the `knitr::kable()` function.
You can find a comment about community guidelines in README.md and CONDUCT.md in the file list.
Ref: https://github.com/KateHyoung/UTDEventData/commit/432c713fdd13a6938c978496de54b157f245f5f0#diff-46b8dc461fb4104140b72e4043aaeefe

README was restructured and now includes more information for users. Please see [here]( https://github.com/KateHyoung/UTDEventData). If you think it needs more information, please let me know.

Package’s vignette is edited and restructured according to the comments of @briatte and @andrewheiss.
Ref. https://github.com/KateHyoung/UTDEventData/blob/master/vignettes/UTDEventData.Rmd

I am trying to list up all my changes from the reviewer’s comments and hope not to miss any important feedback. If I have, please feel free let me know.

This entire review process was very useful to make the package stable and functional. I sincerely appreciate your time and comments for this review.

Best regards,
Kate
