### 2019-04-21 #1322 @alexhanna
@KateHyoung: thanks so much for incorporating these edits. @andrewheiss and @briatte, thanks for the comprehensive comments. Can you two review the changes on the updated package and let me know if you're willing to sign off on the package for meeting the JOSS criteria for acceptance?
