### 2019-04-22 #1322 @briatte
The changes look very thorough indeed, and the updated package loads fine, as does the vignette: I'm signing off too.
