### 2019-04-22 #1322 @alexhanna
You'll need to get a DOI for your repository, @KateHyoung. You'll need to create an archive (on Zenodo, figshare, or somewhere else, if you haven't already) and post the archive DOI here.
