### 2019-04-22 #1322 @danielskatz
Hi - in the future, please notify @openjournals/joss-eics to get the AEiC on duty (me this week)
