### 2019-04-22 #1322 @danielskatz
@KateHyoung - before we accept this, please fix the references in the paper.

1. Items that should be in upper case may need additional {}s around them to preserve the case.
1. For the Salam reference, some additional detail about the paper is needed.  "In. EECS." is not sufficient for someone else to find this.

Once you've done this, please check the resulting pdf by entering ```@whedon generate pdf``` in a new comment here.

Then we'll be ready to accept.
