+++ paper.bib
@@ -1,7 +1,7 @@
 @inproceedings{sayeed16,
   title={Near real-time atrocity event coding},
-  author={Solaimani, Mohiuddin and Salam, Sayeed and Mustafa, Ahmad M and Khan, Latifur and Brandt, Patrick T and Thuraisingham, Bhavani},
-  booktitle={Intelligence and Security Informatics (ISI), 2016 IEEE Conference on},
+  author={Solaimani, Mohiuddin and Salam, Sayeed and Mustafa, Ahmad M and Khan, Latifur and Brandt, Patrick T. and Thuraisingham, Bhavani},
+  booktitle={Intelligence and Security Informatics ({ISI}), 2016 {IEEE} Conference on},
   pages={139--144},
   year={2016},
   organization={IEEE},
@@ -12,9 +12,9 @@
 @inproceedings{brandt2018,
   title={Distributed Framework for Political Event Coding in Real-Time},
   author={Salam, Sayeed and Brandt, Patrick T. and Holmes, Jennifer and Khan, Latifur},
-  booktitle={2nd European Conference on Electrical Engineering & Computer Science (EECS), Bern, Switzerland},
+  booktitle={2nd European Conference on Electrical Engineering & Computer Science ({EECS}), Bern, Switzerland},
   year={2018},
-  organization={EECS}
+  organization={Electrical Engineering & Computer Science}
 }
 
 @Manual{icews,
@@ -30,17 +30,17 @@
  title={Cline Center Historical Phoenix Event Data. v.1.0.0.},
  author={Althaus, Scott and Bajjalieh, Joseph and Carter, John F. and Peyton, Buddy and Shalmon, Dan  A.},
  year={2017},
- note={Distributed by Cline Center for Advanced Social Research.},
+ note={Distributed by Cline Center for Advanced Social Research},
  url = {http://www.clinecenter.illinois.edu/data/event/phoenix/}
 }
 
 @inproceedings{vito2018,
   title={TwoRavens for Event Data},
   author={D'Orazio, Vito and Deng, Marcus and Shoemate, Michael},
-  booktitle={2018 IEEE International Conference on Information Reuse and Integration (IRI)},
+  booktitle={2018 {IEEE} International Conference on Information Reuse and Integration ({IRI})},
   pages={394--401},
   year={2018},
-  organization={IEEE},
+  organization={{IEEE}},
   url={https://doi.org/10.1109/iri.2018.00065},
   doi={10.1109/iri.2018.00065}
 }
