+++ paper.bib
@@ -14,7 +14,7 @@
   author={Salam, Sayeed and Brandt, Patrick T. and Holmes, Jennifer and Khan, Latifur},
   booktitle={2nd European Conference on Electrical Engineering & Computer Science ({EECS}), Bern, Switzerland},
   year={2018},
-  organization={Electrical Engineering & Computer Science}
+  organization={{Electrical Engineering and Computer Science}}
 }
 
 @Manual{icews,
