+++ paper.bib
@@ -1,7 +1,7 @@
 @inproceedings{sayeed16,
   title={Near real-time atrocity event coding},
   author={Solaimani, Mohiuddin and Salam, Sayeed and Mustafa, Ahmad M and Khan, Latifur and Brandt, Patrick T. and Thuraisingham, Bhavani},
-  booktitle={Intelligence and Security Informatics ({ISI}), 2016 {IEEE} Conference on},
+  booktitle={{Intelligence and Security Informatics (ISI), 2016 IEEE Conference}},
   pages={139--144},
   year={2016},
   organization={IEEE},
@@ -12,9 +12,9 @@
 @inproceedings{brandt2018,
   title={Distributed Framework for Political Event Coding in Real-Time},
   author={Salam, Sayeed and Brandt, Patrick T. and Holmes, Jennifer and Khan, Latifur},
-  booktitle={2nd European Conference on Electrical Engineering & Computer Science ({EECS}), Bern, Switzerland},
+  booktitle={{2nd European Conference on Electrical Engineering & Computer Science (EECS), Bern, Switzerland}},
   year={2018},
-  organization={{Electrical Engineering and Computer Science}}
+  organization={EECS}
 }
 
 @Manual{icews,
@@ -37,7 +37,7 @@
 @inproceedings{vito2018,
   title={TwoRavens for Event Data},
   author={D'Orazio, Vito and Deng, Marcus and Shoemate, Michael},
-  booktitle={2018 {IEEE} International Conference on Information Reuse and Integration ({IRI})},
+  booktitle={{2018 IEEE International Conference on Information Reuse and Integration (IRI)}},
   pages={394--401},
   year={2018},
   organization={{IEEE}},
@@ -61,7 +61,7 @@
 @inproceedings{halterman2017,
   title={Adaptive scalable pipelines for political event data generation},
   author={Halterman, Andrew and Irvine, Jill and Landis, Manar and Jalla, Phanindra and Liang, Yan and Grant, Christan and Solaimani, Mohiuddin},
-  booktitle={2017 IEEE International Conference on Big Data (Big Data)},
+  booktitle={{2017 IEEE International Conference on Big Data}},
   pages={2879--2883},
   year={2017},
   organization={IEEE},
