+++ paper.bib
@@ -15,7 +15,6 @@
   booktitle={{2nd European Conference on Electrical Engineering & Computer Science (EECS), Bern, Switzerland}},
   year={2018},
   organization={EECS}
-  url = {http://www.eecs-conf.org/}
 }
 
 @Manual{icews,
