### 2019-04-22 239251a
minor changes for JOSS
