+++ paper.bib
@@ -14,7 +14,8 @@
   author={Salam, Sayeed and Brandt, Patrick T. and Holmes, Jennifer and Khan, Latifur},
   booktitle={{2nd European Conference on Electrical Engineering & Computer Science (EECS), Bern, Switzerland}},
   year={2018},
-  organization={{2nd European Conference on Electrical Engineering & Computer Science (EECS), Bern, Switzerland}}
+  organization={{2nd European Conference on Electrical Engineering & Computer Science (EECS), Bern, Switzerland}},
+  url={http://www.eecs-conf.org/}
 }
 
 @Manual{icews,
