+++ paper.bib
@@ -12,9 +12,9 @@
 @inproceedings{brandt2018,
   title={Distributed Framework for Political Event Coding in Real-Time},
   author={Salam, Sayeed and Brandt, Patrick T. and Holmes, Jennifer and Khan, Latifur},
-  booktitle={{2nd European Conference on Electrical Engineering & Computer Science (EECS), Bern, Switzerland}},
+  booktitle={{2nd European Conference on Electrical Engineering and Computer Science (EECS), Bern, Switzerland}},
   year={2018},
-  organization={{2nd European Conference on Electrical Engineering & Computer Science (EECS), Bern, Switzerland}},
+  organization={{2nd European Conference on Electrical Engineering and Computer Science}},
   url={http://www.eecs-conf.org/}
 }
 
