### 2019-04-22 #1322 @KateHyoung
Dear @danielskatz,
I have corrected and added information for the references and checked other parts of the paper as well. My apology for several attempts for compiling. A style `whedon`  generated differs from my internal compiling system.   I think the draft looks good.

I would like to express my gratitude to the editor, @alexhanna, and the reviewers; @briatte and @andrewheiss, again for their times and comments for this R library and its paper.

Kate
