### 2019-04-22 #1322 @danielskatz
Can you also merge https://github.com/KateHyoung/UTDEventData/pull/12 or make similar changes?
