### 2019-04-22 #1322 @KateHyoung
@danielskatz, I have merged the changes in the bib file. Thank you.
